Simplify fixture by inlining in each scenario
