+++ source/databricks/calculation_engine/tests/features/given_an_energy_calculation/when_with_only_one_exchange_grid_loss_and_system_correction_metering_point/test_output.py
@@ -15,14 +15,34 @@ from pathlib import Path
 from typing import Any
 
 import pytest
+from pyspark.sql import SparkSession
 
 from features.utils.assertion import assert_output
+from features.utils.expected_output import ExpectedOutput
+from features.utils.scenario_executor import ScenarioExecutor
+from package.calculation.calculation_results import CalculationResultsContainer
 
 
 def get_output_names() -> list[str]:
     output_folder_path = Path(__file__).parent / "output"
     csv_files = list(output_folder_path.rglob("*.csv"))
     return [Path(file).stem for file in csv_files]
+
+
+@pytest.fixture(scope="module")
+def actual_and_expected(
+    spark: SparkSession,
+) -> tuple[CalculationResultsContainer, list[ExpectedOutput]]:
+    """
+    Provides the actual and expected output for a scenario test case.
+
+    IMPORTANT: It is crucial that this fixture has scope=module, as the scenario executor
+    is specific to a single scenario, which are each located in their own module.
+    """
+
+    scenario_path = str(Path(__file__).parent)
+    scenario_executor = ScenarioExecutor(spark)
+    return scenario_executor.execute(scenario_path)
 
 
 # IMPORTANT:
